chore: create token endpoints file
