+++ components/brave_wallet_ui/common/slices/endpoints/nfts.endpoints.ts
@@ -54,25 +54,10 @@ export const nftsEndpoints = ({
     >({
       queryFn: async (tokenArg, { endpoint }, _extraOptions, baseQuery) => {
         try {
-          if (!tokenArg.isErc721) {
-            throw new Error(
-              'Cannot fetch erc-721 metadata for non erc-721 token'
-            )
-          }
+          const { getErc721Metadata } = baseQuery(undefined).cache
 
-          const { jsonRpcService } = baseQuery(undefined).data
+          const metadata: ERC721Metadata = await getErc721Metadata(tokenArg)
 
-          const result = await jsonRpcService.getERC721Metadata(
-            tokenArg.contractAddress,
-            tokenArg.tokenId,
-            tokenArg.chainId
-          )
-
-          if (result.error || result.errorMessage) {
-            throw new Error(result.errorMessage)
-          }
-
-          const metadata: ERC721Metadata = JSON.parse(result.response)
           return {
             data: {
               id: blockchainTokenEntityAdaptor.selectId(tokenArg),
